refactor(ai): modularize MCP tools adapter with provider-specific schema sanitization
- Extract schema sanitization into dedicated schemaSanitizer module with provider registry
  - geminiSanitizer: strict validation for Google Generative AI compatibility
  - baseSanitizer: passthrough for permissive providers (OpenAI, Anthropic)
  - Registry pattern allows easy addition of new provider-specific sanitizers

- Extract widget rendering into dedicated widgets module
  - htmlUtils: escapeHtml, injectDataIntoHtml utilities
  - skybridgeBridge: OpenAI/Skybridge widget bridge injection
  - templates/: weatherCard, calculator, textAnalysis, generic widgets

- Reduce mcpToolsAdapter.ts from 1479 to 500 lines (-66%)
- Fix Gemini API validation errors for bare object schemas and empty items
- Update AI SDK dependencies to latest versions
