+++ package.json
@@ -30,11 +30,11 @@
   },
   "dependencies": {
     "@ai-sdk/anthropic": "^2.0.56",
-    "@ai-sdk/gateway": "^2.0.21",
-    "@ai-sdk/google": "^2.0.46",
-    "@ai-sdk/openai": "^2.0.85",
+    "@ai-sdk/gateway": "^2.0.22",
+    "@ai-sdk/google": "^2.0.49",
+    "@ai-sdk/openai": "^2.0.88",
     "@ai-sdk/openai-compatible": "^1.0.26",
-    "@ai-sdk/react": "^2.0.114",
+    "@ai-sdk/react": "^2.0.117",
     "@ai-sdk/ui-utils": "^1.2.11",
     "@codemirror/lang-json": "^6.0.2",
     "@codemirror/theme-one-dark": "^6.1.3",
@@ -42,9 +42,9 @@
     "@huggingface/inference": "^4.13.5",
     "@libsql/client": "^0.15.12",
     "@mcp-ui/client": "^5.17.1",
-    "@mcp-ui/server": "^5.16.2",
+    "@mcp-ui/server": "^5.16.3",
     "@modelcontextprotocol/sdk": "^1.24.3",
-    "@openrouter/ai-sdk-provider": "^1.5.3",
+    "@openrouter/ai-sdk-provider": "^1.5.4",
     "@radix-ui/react-accordion": "^1.2.12",
     "@radix-ui/react-alert-dialog": "^1.1.15",
     "@radix-ui/react-aspect-ratio": "^1.1.7",
@@ -77,7 +77,7 @@
     "@supabase/supabase-js": "^2.87.1",
     "@tailwindcss/vite": "^4.1.17",
     "@uiw/react-codemirror": "^4.25.3",
-    "ai": "^5.0.112",
+    "ai": "^5.0.115",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
@@ -92,7 +92,7 @@
     "katex": "^0.16.27",
     "keytar": "^7.9.0",
     "lucide-react": "^0.541.0",
-    "mcp-use": "^1.10.0",
+    "mcp-use": "^1.11.2",
     "mermaid": "^11.12.2",
     "next-themes": "^0.4.6",
     "pdf-parse": "^2.4.5",
