+++ src/main/services/ai/mcpToolsAdapter.ts
@@ -1,698 +1,25 @@
-import { tool } from "ai";
-import { z } from "zod/v3";
+/**
+ * MCP Tools Adapter
+ *
+ * Converts MCP (Model Context Protocol) tools to AI SDK format.
+ * Handles schema sanitization for provider compatibility and widget rendering.
+ */
+
+import { tool, jsonSchema } from "ai";
 import { mcpService, configManager } from "../../ipc/mcpHandlers";
 import { mcpHealthService } from "../mcpHealthService";
 import type { Tool } from "../../types/mcp";
 import { getLogger } from '../logging';
 
+// Import from new modules
+import { sanitizeSchema } from './schemaSanitizer';
+import {
+  generateWidgetHtml,
+  injectDataIntoHtml,
+  injectSkybridgeBridge,
+} from './widgets';
+
 const logger = getLogger();
-
-/**
- * Generate HTML for an mcp-use widget
- * Creates a self-contained HTML document that renders the widget
- */
-function generateWidgetHtml(
-  widgetMeta: { name: string; description?: string; type: string },
-  data: Record<string, any>
-): string {
-  const { name } = widgetMeta;
-
-  // Generate widget-specific HTML based on the widget name
-  switch (name) {
-    case 'weather-card':
-      return generateWeatherCardHtml(data);
-    case 'calculator-result':
-      return generateCalculatorResultHtml(data);
-    case 'text-analysis-chart':
-      return generateTextAnalysisChartHtml(data);
-    default:
-      // Generic widget fallback
-      return generateGenericWidgetHtml(name, data);
-  }
-}
-
-/**
- * Weather card widget HTML
- */
-function generateWeatherCardHtml(data: Record<string, any>): string {
-  const { city, condition, temperature, unit, humidity, timestamp } = data;
-
-  return `<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <style>
-    * { margin: 0; padding: 0; box-sizing: border-box; }
-    body {
-      font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
-      background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-      min-height: 100vh;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      padding: 20px;
-    }
-    .card {
-      background: rgba(255, 255, 255, 0.95);
-      border-radius: 20px;
-      padding: 30px;
-      width: 100%;
-      max-width: 350px;
-      box-shadow: 0 10px 40px rgba(0, 0, 0, 0.2);
-    }
-    .city {
-      font-size: 24px;
-      font-weight: 600;
-      color: #333;
-      margin-bottom: 5px;
-    }
-    .timestamp {
-      font-size: 12px;
-      color: #888;
-      margin-bottom: 20px;
-    }
-    .temp-container {
-      display: flex;
-      align-items: center;
-      margin-bottom: 20px;
-    }
-    .temp {
-      font-size: 64px;
-      font-weight: 300;
-      color: #333;
-    }
-    .unit {
-      font-size: 24px;
-      color: #666;
-      margin-left: 5px;
-      align-self: flex-start;
-      margin-top: 10px;
-    }
-    .condition {
-      font-size: 18px;
-      color: #666;
-      text-transform: capitalize;
-      margin-bottom: 20px;
-    }
-    .details {
-      display: flex;
-      gap: 20px;
-      padding-top: 20px;
-      border-top: 1px solid #eee;
-    }
-    .detail-item {
-      text-align: center;
-    }
-    .detail-label {
-      font-size: 12px;
-      color: #888;
-      margin-bottom: 5px;
-    }
-    .detail-value {
-      font-size: 16px;
-      font-weight: 500;
-      color: #333;
-    }
-    .weather-icon {
-      font-size: 48px;
-      margin-left: auto;
-    }
-  </style>
-</head>
-<body>
-  <div class="card">
-    <div class="city">${escapeHtml(city || 'Unknown')}</div>
-    <div class="timestamp">${escapeHtml(timestamp || new Date().toLocaleString())}</div>
-    <div class="temp-container">
-      <span class="temp">${escapeHtml(String(temperature ?? '--'))}</span>
-      <span class="unit">${escapeHtml(unit || '°C')}</span>
-      <span class="weather-icon">${getWeatherEmoji(condition)}</span>
-    </div>
-    <div class="condition">${escapeHtml(condition || 'Unknown')}</div>
-    <div class="details">
-      <div class="detail-item">
-        <div class="detail-label">Humidity</div>
-        <div class="detail-value">${escapeHtml(String(humidity ?? '--'))}${typeof humidity === 'number' ? '%' : ''}</div>
-      </div>
-    </div>
-  </div>
-</body>
-</html>`;
-}
-
-/**
- * Calculator result widget HTML
- */
-function generateCalculatorResultHtml(data: Record<string, any>): string {
-  const { expression, result } = data;
-
-  return `<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <style>
-    * { margin: 0; padding: 0; box-sizing: border-box; }
-    body {
-      font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
-      background: #1a1a2e;
-      min-height: 100vh;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      padding: 20px;
-    }
-    .calculator {
-      background: #16213e;
-      border-radius: 15px;
-      padding: 25px;
-      min-width: 280px;
-      box-shadow: 0 10px 30px rgba(0, 0, 0, 0.3);
-    }
-    .expression {
-      font-size: 18px;
-      color: #888;
-      text-align: right;
-      margin-bottom: 10px;
-      font-family: 'SF Mono', Monaco, monospace;
-    }
-    .result {
-      font-size: 42px;
-      font-weight: 300;
-      color: #fff;
-      text-align: right;
-      font-family: 'SF Mono', Monaco, monospace;
-    }
-  </style>
-</head>
-<body>
-  <div class="calculator">
-    <div class="expression">${escapeHtml(expression || '')}</div>
-    <div class="result">= ${escapeHtml(String(result ?? ''))}</div>
-  </div>
-</body>
-</html>`;
-}
-
-/**
- * Text analysis chart widget HTML
- */
-function generateTextAnalysisChartHtml(data: Record<string, any>): string {
-  const { wordCount, charCount, sentenceCount, avgWordLength } = data;
-
-  return `<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <style>
-    * { margin: 0; padding: 0; box-sizing: border-box; }
-    body {
-      font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
-      background: #f5f5f5;
-      min-height: 100vh;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      padding: 20px;
-    }
-    .chart {
-      background: #fff;
-      border-radius: 15px;
-      padding: 25px;
-      width: 100%;
-      max-width: 400px;
-      box-shadow: 0 5px 20px rgba(0, 0, 0, 0.1);
-    }
-    .title {
-      font-size: 18px;
-      font-weight: 600;
-      color: #333;
-      margin-bottom: 20px;
-    }
-    .stat {
-      display: flex;
-      justify-content: space-between;
-      padding: 12px 0;
-      border-bottom: 1px solid #eee;
-    }
-    .stat:last-child { border-bottom: none; }
-    .stat-label { color: #666; }
-    .stat-value { font-weight: 600; color: #333; }
-  </style>
-</head>
-<body>
-  <div class="chart">
-    <div class="title">Text Analysis</div>
-    <div class="stat">
-      <span class="stat-label">Words</span>
-      <span class="stat-value">${escapeHtml(String(wordCount ?? 0))}</span>
-    </div>
-    <div class="stat">
-      <span class="stat-label">Characters</span>
-      <span class="stat-value">${escapeHtml(String(charCount ?? 0))}</span>
-    </div>
-    <div class="stat">
-      <span class="stat-label">Sentences</span>
-      <span class="stat-value">${escapeHtml(String(sentenceCount ?? 0))}</span>
-    </div>
-    <div class="stat">
-      <span class="stat-label">Avg. Word Length</span>
-      <span class="stat-value">${escapeHtml(String(avgWordLength ?? 0))}</span>
-    </div>
-  </div>
-</body>
-</html>`;
-}
-
-/**
- * Generic widget fallback HTML
- */
-function generateGenericWidgetHtml(name: string, data: Record<string, any>): string {
-  const dataEntries = Object.entries(data)
-    .filter(([key]) => !key.startsWith('_'))
-    .map(([key, value]) => `
-      <div class="item">
-        <span class="key">${escapeHtml(key)}</span>
-        <span class="value">${escapeHtml(String(value))}</span>
-      </div>
-    `).join('');
-
-  return `<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <style>
-    * { margin: 0; padding: 0; box-sizing: border-box; }
-    body {
-      font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
-      background: #f8f9fa;
-      min-height: 100vh;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      padding: 20px;
-    }
-    .widget {
-      background: #fff;
-      border-radius: 12px;
-      padding: 20px;
-      width: 100%;
-      max-width: 350px;
-      box-shadow: 0 4px 15px rgba(0, 0, 0, 0.1);
-    }
-    .title {
-      font-size: 16px;
-      font-weight: 600;
-      color: #333;
-      margin-bottom: 15px;
-      text-transform: capitalize;
-    }
-    .item {
-      display: flex;
-      justify-content: space-between;
-      padding: 8px 0;
-      border-bottom: 1px solid #eee;
-    }
-    .item:last-child { border-bottom: none; }
-    .key { color: #666; text-transform: capitalize; }
-    .value { font-weight: 500; color: #333; }
-  </style>
-</head>
-<body>
-  <div class="widget">
-    <div class="title">${escapeHtml(name.replace(/-/g, ' '))}</div>
-    ${dataEntries}
-  </div>
-</body>
-</html>`;
-}
-
-/**
- * Escape HTML special characters
- */
-function escapeHtml(str: string): string {
-  return str
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/"/g, '&quot;')
-    .replace(/'/g, '&#039;');
-}
-
-/**
- * Inject data into widget HTML template
- * Adds a script that:
- * 1. Sets window variables for immediate access (fallback)
- * 2. Implements @mcp-ui/client postMessage protocol to receive iframeRenderData
- *
- * The postMessage protocol:
- * - Parent sends 'ui-lifecycle-iframe-render-data' with payload.renderData
- * - Iframe notifies parent with 'ui-lifecycle-iframe-ready' when ready
- */
-function injectDataIntoHtml(html: string, data: Record<string, any>): string {
-  // Escape data for safe inclusion in script
-  const jsonData = JSON.stringify(data);
-
-  const dataScript = `<script>
-    // Fallback: Set data immediately for templates that read from window variables
-    window.__data = ${jsonData};
-    window.__props = ${jsonData};
-    window.__widgetData = ${jsonData};
-    window.__IFRAME_RENDER_DATA__ = ${jsonData};
-
-    // mcp-use Apps SDK format: window.openai with toolInput and toolOutput
-    // useWidget hook reads props from toolInput, output from toolOutput
-    window.openai = window.openai || {};
-    window.openai.toolInput = ${jsonData};
-    window.openai.toolOutput = ${jsonData};
-
-    // @mcp-ui/client postMessage protocol for receiving iframeRenderData
-    (function() {
-      var renderDataReceived = false;
-
-      // Listen for render data from parent (UIResourceRenderer)
-      window.addEventListener('message', function(event) {
-        if (event.data && event.data.type === 'ui-lifecycle-iframe-render-data') {
-          if (renderDataReceived) return;
-          renderDataReceived = true;
-
-          var renderData = event.data.payload && event.data.payload.renderData;
-          if (renderData) {
-            // Update window variables with data from parent
-            window.__data = renderData;
-            window.__props = renderData;
-            window.__widgetData = renderData;
-            window.__IFRAME_RENDER_DATA__ = renderData;
-
-            // mcp-use Apps SDK format
-            window.openai = window.openai || {};
-            window.openai.toolInput = renderData;
-            window.openai.toolOutput = renderData;
-
-            // Dispatch custom event for templates that listen for data
-            window.dispatchEvent(new CustomEvent('mcp-render-data', { detail: renderData }));
-
-            // If template has a renderUI function, call it
-            if (typeof window.renderUI === 'function') {
-              window.renderUI(renderData);
-            }
-
-            // If template has updateUI function, call it
-            if (typeof window.updateUI === 'function') {
-              window.updateUI(renderData);
-            }
-          }
-        }
-      });
-
-      // Notify parent that iframe is ready to receive data
-      if (window.parent && window.parent !== window) {
-        window.parent.postMessage({ type: 'ui-lifecycle-iframe-ready' }, '*');
-      }
-
-      // Auto-resize: Send height to parent using @mcp-ui/client protocol
-      // Uses 'ui-size-change' message type as per mcp-ui documentation
-      function sendHeight() {
-        if (window.parent && window.parent !== window) {
-          var height = Math.max(
-            document.body.scrollHeight,
-            document.body.offsetHeight,
-            document.documentElement.scrollHeight,
-            document.documentElement.offsetHeight
-          );
-          // Minimum height of 200px, maximum of 600px
-          height = Math.max(200, Math.min(height, 600));
-          window.parent.postMessage({
-            type: 'ui-size-change',
-            payload: { height: height }
-          }, '*');
-        }
-      }
-
-      // Use ResizeObserver for accurate height tracking (recommended by mcp-ui)
-      if (typeof ResizeObserver !== 'undefined') {
-        var resizeObserver = new ResizeObserver(function(entries) {
-          entries.forEach(function(entry) {
-            var height = Math.max(200, Math.min(entry.contentRect.height, 600));
-            window.parent.postMessage({
-              type: 'ui-size-change',
-              payload: { height: height }
-            }, '*');
-          });
-        });
-        resizeObserver.observe(document.documentElement);
-      }
-
-      // Fallback: Send height on load
-      if (document.readyState === 'loading') {
-        document.addEventListener('DOMContentLoaded', function() {
-          setTimeout(sendHeight, 100);
-        });
-      } else {
-        setTimeout(sendHeight, 100);
-      }
-
-      // Also send height after window load (for images, etc.)
-      window.addEventListener('load', function() {
-        setTimeout(sendHeight, 200);
-      });
-    })();
-  </script>`;
-
-  // Try to inject after <head> tag
-  if (html.includes('<head>')) {
-    return html.replace('<head>', `<head>\n${dataScript}`);
-  }
-  // Try to inject after <html> tag
-  if (html.includes('<html')) {
-    return html.replace(/<html[^>]*>/, `$&\n<head>${dataScript}</head>`);
-  }
-  // Fallback: prepend to HTML
-  return dataScript + html;
-}
-
-/**
- * Inject Skybridge/OpenAI compatible bridge into widget HTML
- * Creates a full window.openai API that Skybridge widgets expect:
- * - Globals: locale, theme, displayMode, toolInput, toolOutput, toolResponseMetadata
- * - Methods: callTool, sendFollowUpMessage, requestDisplayMode, openExternal, setWidgetState
- *
- * Methods communicate with parent via postMessage using 'openai-bridge-*' message types
- */
-function injectSkybridgeBridge(
-  html: string,
-  options: {
-    toolInput: Record<string, any>;
-    toolOutput: Record<string, any>;
-    responseMetadata: Record<string, any>;
-    locale?: string;
-  }
-): string {
-  const { toolInput, toolOutput, responseMetadata, locale = 'en-US' } = options;
-
-  const bridgeScript = `<script>
-    // Skybridge/OpenAI Widget Bridge for Levante
-    (function() {
-      var pendingCallbacks = {};
-      var callbackId = 0;
-      var widgetState = {};
-
-      // Initialize window.openai with globals and methods
-      window.openai = {
-        // Globals (read-only data)
-        locale: ${JSON.stringify(locale)},
-        theme: 'light',
-        displayMode: 'inline',
-        maxHeight: 600,
-        safeArea: { top: 0, right: 0, bottom: 0, left: 0 },
-        toolInput: ${JSON.stringify(toolInput)},
-        toolOutput: ${JSON.stringify(toolOutput)},
-        toolResponseMetadata: ${JSON.stringify(responseMetadata)},
-        widgetState: widgetState,
-
-        // Methods
-        callTool: function(toolName, args) {
-          return new Promise(function(resolve, reject) {
-            var id = ++callbackId;
-            pendingCallbacks[id] = { resolve: resolve, reject: reject };
-            window.parent.postMessage({
-              type: 'openai-bridge-call-tool',
-              payload: { id: id, toolName: toolName, args: args || {} }
-            }, '*');
-          });
-        },
-
-        sendFollowUpMessage: function(message) {
-          window.parent.postMessage({
-            type: 'openai-bridge-follow-up',
-            payload: { message: message }
-          }, '*');
-        },
-
-        requestDisplayMode: function(options) {
-          var mode = typeof options === 'string' ? options : (options && options.mode) || 'inline';
-          window.openai.displayMode = mode;
-          window.parent.postMessage({
-            type: 'openai-bridge-display-mode',
-            payload: { mode: mode }
-          }, '*');
-        },
-
-        openExternal: function(url) {
-          window.parent.postMessage({
-            type: 'openai-bridge-open-external',
-            payload: { url: url }
-          }, '*');
-        },
-
-        requestClose: function() {
-          window.parent.postMessage({
-            type: 'openai-bridge-close'
-          }, '*');
-        },
-
-        setWidgetState: function(state) {
-          widgetState = state;
-          window.openai.widgetState = state;
-          window.parent.postMessage({
-            type: 'openai-bridge-set-state',
-            payload: { state: state }
-          }, '*');
-        }
-      };
-
-      // Listen for responses from parent
-      window.addEventListener('message', function(event) {
-        var data = event.data;
-        if (!data || !data.type) return;
-
-        // Handle callTool response
-        if (data.type === 'openai-bridge-call-tool-response') {
-          var callback = pendingCallbacks[data.payload.id];
-          if (callback) {
-            delete pendingCallbacks[data.payload.id];
-            if (data.payload.error) {
-              callback.reject(new Error(data.payload.error));
-            } else {
-              callback.resolve(data.payload.result);
-            }
-          }
-        }
-
-        // Handle globals update from parent
-        if (data.type === 'openai-bridge-set-globals' || data.type === 'openai:set_globals') {
-          var globals = data.payload || data.globals || {};
-          if (globals.theme) window.openai.theme = globals.theme;
-          if (globals.locale) window.openai.locale = globals.locale;
-          if (globals.displayMode) window.openai.displayMode = globals.displayMode;
-          if (globals.maxHeight) window.openai.maxHeight = globals.maxHeight;
-          if (globals.toolInput) window.openai.toolInput = globals.toolInput;
-          if (globals.toolOutput) window.openai.toolOutput = globals.toolOutput;
-          if (globals.toolResponseMetadata) window.openai.toolResponseMetadata = globals.toolResponseMetadata;
-
-          // Dispatch event for hooks that listen to globals changes
-          window.dispatchEvent(new CustomEvent('openai:set_globals', { detail: globals }));
-        }
-
-        // Handle iframeRenderData from @mcp-ui/client
-        if (data.type === 'ui-lifecycle-iframe-render-data') {
-          var renderData = data.payload && data.payload.renderData;
-          if (renderData) {
-            // Update toolInput/toolOutput with render data
-            window.openai.toolInput = Object.assign({}, window.openai.toolInput, renderData);
-            window.openai.toolOutput = Object.assign({}, window.openai.toolOutput, renderData);
-            if (renderData.theme) window.openai.theme = renderData.theme;
-            if (renderData.locale) window.openai.locale = renderData.locale;
-
-            // Also set legacy variables
-            window.__IFRAME_RENDER_DATA__ = renderData;
-            window.__data = renderData;
-            window.__props = renderData;
-          }
-        }
-      });
-
-      // Auto-resize: Send height to parent
-      function sendHeight() {
-        if (window.parent && window.parent !== window) {
-          var height = Math.max(
-            document.body.scrollHeight,
-            document.body.offsetHeight,
-            document.documentElement.scrollHeight,
-            document.documentElement.offsetHeight
-          );
-          height = Math.max(200, Math.min(height, 600));
-          window.parent.postMessage({
-            type: 'ui-size-change',
-            payload: { height: height }
-          }, '*');
-        }
-      }
-
-      // Use ResizeObserver for accurate height tracking
-      if (typeof ResizeObserver !== 'undefined') {
-        var resizeObserver = new ResizeObserver(function(entries) {
-          entries.forEach(function(entry) {
-            var height = Math.max(200, Math.min(entry.contentRect.height, 600));
-            window.parent.postMessage({
-              type: 'ui-size-change',
-              payload: { height: height }
-            }, '*');
-          });
-        });
-        resizeObserver.observe(document.documentElement);
-      }
-
-      // Notify parent that iframe is ready
-      if (window.parent && window.parent !== window) {
-        window.parent.postMessage({ type: 'ui-lifecycle-iframe-ready' }, '*');
-      }
-
-      // Send initial height
-      if (document.readyState === 'loading') {
-        document.addEventListener('DOMContentLoaded', function() {
-          setTimeout(sendHeight, 100);
-        });
-      } else {
-        setTimeout(sendHeight, 100);
-      }
-
-      window.addEventListener('load', function() {
-        setTimeout(sendHeight, 200);
-      });
-    })();
-  </script>`;
-
-  // Inject as first script in head to ensure window.openai is available before widget code
-  if (html.includes('<head>')) {
-    return html.replace('<head>', `<head>\n${bridgeScript}`);
-  }
-  if (html.includes('<head ')) {
-    return html.replace(/<head[^>]*>/, `$&\n${bridgeScript}`);
-  }
-  // Try to inject after <html> tag
-  if (html.includes('<html')) {
-    return html.replace(/<html[^>]*>/, `$&\n<head>${bridgeScript}</head>`);
-  }
-  // Fallback: prepend to HTML
-  return bridgeScript + html;
-}
-
-/**
- * Get weather emoji based on condition
- */
-function getWeatherEmoji(condition: string): string {
-  const c = (condition || '').toLowerCase();
-  if (c.includes('sun') || c.includes('clear')) return '☀️';
-  if (c.includes('cloud') || c.includes('overcast')) return '☁️';
-  if (c.includes('partly')) return '⛅';
-  if (c.includes('rain') || c.includes('shower')) return '🌧️';
-  if (c.includes('thunder') || c.includes('storm')) return '⛈️';
-  if (c.includes('snow')) return '❄️';
-  if (c.includes('fog') || c.includes('mist')) return '🌫️';
-  if (c.includes('wind')) return '💨';
-  return '🌤️';
-}
 
 /**
  * Get all MCP tools from connected servers and convert them to AI SDK format
@@ -837,147 +164,10 @@ export async function getMCPTools(): Promise<Record<string, any>> {
 }
 
 /**
- * Recursively convert JSON Schema to Zod schema
- * Preserves array items definitions for proper Google Generative AI compatibility
- */
-function convertJsonSchemaToZod(schema: any, debugPath?: string): any {
-  if (!schema || typeof schema !== 'object') {
-    return z.any().describe(schema?.description || "");
-  }
-
-  const description = schema.description || "";
-
-  // Handle array type with items
-  if (schema.type === "array") {
-    // Log problematic arrays for debugging
-    if (debugPath && (debugPath.includes('nodes') || debugPath.includes('operations'))) {
-      logger.aiSdk.debug("Converting array schema", {
-        path: debugPath,
-        hasItems: !!schema.items,
-        itemsType: typeof schema.items,
-        itemsValue: JSON.stringify(schema.items, null, 2),
-        fullSchema: JSON.stringify(schema, null, 2)
-      });
-    }
-
-    if (schema.items) {
-      // Handle array of schemas (tuple type) - not commonly used but possible
-      if (Array.isArray(schema.items)) {
-        logger.aiSdk.warn("Array items is an array (tuple type), using any[]", {
-          path: debugPath,
-          itemsLength: schema.items.length
-        });
-        return z.array(z.any()).describe(description);
-      }
-
-      // Recursively convert items schema
-      const itemsSchema = convertJsonSchemaToZod(schema.items, `${debugPath || ''}.items`);
-      return z.array(itemsSchema).describe(description);
-    }
-
-    // Fallback if no items defined
-    logger.aiSdk.warn("Array without items definition", {
-      path: debugPath,
-      schema: JSON.stringify(schema, null, 2)
-    });
-    return z.array(z.any()).describe(description);
-  }
-
-  // Handle object type
-  if (schema.type === "object") {
-    // Object with specific properties
-    if (schema.properties && Object.keys(schema.properties).length > 0) {
-      const schemaObj: Record<string, any> = {};
-      const propertyNames = Object.keys(schema.properties);
-
-      // Validate required array - filter out properties that don't exist
-      // This is crucial for Gemini API which strictly validates this
-      const validRequired = schema.required
-        ? (schema.required as string[]).filter(reqProp => {
-            const exists = propertyNames.includes(reqProp);
-            if (!exists) {
-              logger.aiSdk.warn("Required property not found in properties (Gemini will reject this)", {
-                path: debugPath,
-                requiredProperty: reqProp,
-                availableProperties: propertyNames
-              });
-            }
-            return exists;
-          })
-        : [];
-
-      for (const [propName, propDef] of Object.entries(schema.properties)) {
-        const propSchema = convertJsonSchemaToZod(propDef, `${debugPath || ''}.${propName}`);
-
-        // Handle required fields using validated required array
-        if (validRequired.includes(propName)) {
-          schemaObj[propName] = propSchema;
-        } else {
-          schemaObj[propName] = propSchema.optional();
-        }
-      }
-
-      return z.object(schemaObj).describe(description);
-    }
-
-    // Object with additionalProperties (generic object with dynamic keys)
-    // This handles cases like { "type": "object", "additionalProperties": true }
-    if (schema.additionalProperties === true || (schema.additionalProperties && typeof schema.additionalProperties === 'object')) {
-      logger.aiSdk.debug("Converting object with additionalProperties to z.record()", {
-        path: debugPath,
-        hasAdditionalProperties: !!schema.additionalProperties,
-        additionalPropertiesType: typeof schema.additionalProperties
-      });
-
-      // Use z.record() for objects with dynamic keys
-      // If additionalProperties is an object (schema), convert it recursively
-      if (typeof schema.additionalProperties === 'object') {
-        const valueSchema = convertJsonSchemaToZod(schema.additionalProperties, `${debugPath || ''}.additionalProperties`);
-        return z.record(valueSchema).describe(description);
-      }
-
-      // If additionalProperties is just true, use z.any() for values
-      return z.record(z.any()).describe(description);
-    }
-
-    // IMPORTANT: Object with no properties and no additionalProperties
-    // This represents "any object" in JSON Schema, so use z.record(z.any())
-    // Using z.object({}) would create a closed empty object that Gemini rejects
-    logger.aiSdk.debug("Converting bare object (no properties) to z.record(z.any())", {
-      path: debugPath,
-      reason: "Gemini requires objects without properties to be treated as dynamic records"
-    });
-    return z.record(z.any()).describe(description);
-  }
-
-  // Handle primitive types
-  switch (schema.type) {
-    case "string":
-      return z.string().describe(description);
-    case "number":
-    case "integer":
-      return z.number().describe(description);
-    case "boolean":
-      return z.boolean().describe(description);
-    case "null":
-      return z.null().describe(description);
-    default:
-      return z.any().describe(description);
-  }
-}
-
-/**
  * Convert an MCP tool to AI SDK format
  */
 function createAISDKTool(serverId: string, mcpTool: Tool) {
   logger.aiSdk.debug("Creating AI SDK tool", { serverId, toolName: mcpTool.name });
-
-  // Log the RAW MCP schema BEFORE any conversion
-  logger.aiSdk.debug("Raw MCP tool schema", {
-    serverId,
-    toolName: mcpTool.name,
-    fullInputSchema: JSON.stringify(mcpTool.inputSchema, null, 2)
-  });
 
   // Validate tool name
   if (!mcpTool.name || mcpTool.name.trim() === "") {
@@ -986,63 +176,35 @@ function createAISDKTool(serverId: string, mcpTool: Tool) {
     );
   }
 
-  // Create a schema from MCP tool input schema
-  let inputSchema = z.object({});
+  // Sanitize MCP JSON Schema for provider compatibility
+  // Uses the strictest sanitization (Gemini) by default for maximum compatibility
+  let inputSchema: ReturnType<typeof jsonSchema>;
 
   try {
-    if (mcpTool.inputSchema && mcpTool.inputSchema.properties) {
-      const schemaObj: Record<string, any> = {};
-      const propertyNames = Object.keys(mcpTool.inputSchema.properties);
+    if (mcpTool.inputSchema) {
+      // Sanitize schema: fix objects without properties, arrays without items,
+      // and filter invalid required references
+      const sanitizedSchema = sanitizeSchema(mcpTool.inputSchema, undefined, mcpTool.name);
 
-      // Log the full schema for debugging Gemini issues
-      logger.aiSdk.debug("Processing tool schema", {
+      logger.aiSdk.debug("Sanitized MCP schema", {
         toolName: mcpTool.name,
         serverId,
-        hasRequired: !!mcpTool.inputSchema.required,
-        requiredArray: mcpTool.inputSchema.required || [],
-        propertyNames,
-        fullSchema: JSON.stringify(mcpTool.inputSchema, null, 2)
+        originalType: mcpTool.inputSchema.type,
+        sanitizedType: sanitizedSchema.type,
+        hasProperties: !!sanitizedSchema.properties
       });
 
-      // Validate required array - filter out properties that don't exist
-      // This is crucial for Gemini API which strictly validates this
-      const validRequired = mcpTool.inputSchema.required
-        ? (mcpTool.inputSchema.required as string[]).filter(reqProp => {
-            const exists = propertyNames.includes(reqProp);
-            if (!exists) {
-              logger.aiSdk.warn("Tool has required property not in properties (Gemini will reject)", {
-                toolName: mcpTool.name,
-                requiredProperty: reqProp,
-                availableProperties: propertyNames,
-                serverId
-              });
-            }
-            return exists;
-          })
-        : [];
-
-      for (const [propName, propDef] of Object.entries(
-        mcpTool.inputSchema.properties
-      )) {
-        const propInfo = propDef as any;
-
-        // Use recursive converter to properly handle nested structures
-        // Pass tool name and property for better debugging
-        schemaObj[propName] = convertJsonSchemaToZod(propInfo, `${mcpTool.name}.${propName}`);
-
-        // Handle required fields using validated required array
-        if (!validRequired.includes(propName)) {
-          schemaObj[propName] = schemaObj[propName].optional();
-        }
-      }
-
-      inputSchema = z.object(schemaObj);
+      inputSchema = jsonSchema(sanitizedSchema);
+    } else {
+      // Fallback to empty object schema
+      inputSchema = jsonSchema({ type: "object", properties: {} });
     }
   } catch (error) {
-    logger.aiSdk.warn("Failed to parse schema for tool", {
+    logger.aiSdk.warn("Failed to sanitize schema for tool, using fallback", {
       toolName: mcpTool.name,
       error
     });
+    inputSchema = jsonSchema({ type: "object", properties: {} });
   }
 
   const aiTool = tool({
@@ -1061,17 +223,13 @@ function createAISDKTool(serverId: string, mcpTool: Tool) {
           arguments: args,
         });
 
-        logger.aiSdk.info("[AI-SDK] Raw MCP tool result", {
+        logger.aiSdk.debug("[AI-SDK] Raw MCP tool result", {
           toolName: mcpTool.name,
           serverId,
           contentLength: result.content?.length || 0,
           hasMeta: !!result._meta,
           hasWidgetMeta: !!result._meta?.['mcp-use/widget'],
-          hasOpenAIOutputTemplateInResult: !!result._meta?.['openai/outputTemplate'],
-          hasOpenAIOutputTemplateInTool: !!mcpTool._meta?.['openai/outputTemplate'],
-          toolMetaKeys: mcpTool._meta ? Object.keys(mcpTool._meta) : [],
           hasStructuredContent: !!result.structuredContent,
-          fullResult: JSON.stringify(result, null, 2)
         });
 
         // Check for mcp-use widget in _meta
@@ -1081,302 +239,15 @@ function createAISDKTool(serverId: string, mcpTool: Tool) {
                                       mcpTool._meta?.['openai/outputTemplate'];
 
         if (widgetMeta) {
-          logger.aiSdk.info("[AI-SDK] Found mcp-use widget in _meta", {
-            toolName: mcpTool.name,
-            widgetName: widgetMeta.name,
-            widgetType: widgetMeta.type,
-            hasHtml: !!widgetMeta.html,
-            htmlLength: widgetMeta.html?.length || 0,
-            hasStructuredContent: !!result.structuredContent,
-          });
-
-          // Use HTML from server if available, otherwise generate fallback
-          let widgetHtml: string;
-          const widgetData = result.structuredContent || args;
-
-          if (widgetMeta.html && typeof widgetMeta.html === 'string') {
-            // Server provided the HTML template - inject the data into it
-            widgetHtml = injectDataIntoHtml(widgetMeta.html, widgetData);
-            logger.aiSdk.info("[AI-SDK] Using server-provided HTML for widget with injected data", {
-              widgetName: widgetMeta.name,
-              htmlLength: widgetHtml.length,
-              dataKeys: Object.keys(widgetData),
-            });
-          } else {
-            // Fallback: generate HTML ourselves
-            widgetHtml = generateWidgetHtml(widgetMeta, widgetData);
-            logger.aiSdk.info("[AI-SDK] Generated fallback HTML for widget", {
-              widgetName: widgetMeta.name,
-              htmlLength: widgetHtml.length,
-            });
-          }
-
-          // Create synthetic UI resource with widget data in _meta
-          const uiResource = {
-            type: "resource",
-            resource: {
-              uri: `ui://widget/${widgetMeta.name}.html`,
-              mimeType: "text/html",
-              text: widgetHtml,
-              // Include widget data in _meta for UIResourceRenderer to access
-              _meta: {
-                widgetName: widgetMeta.name,
-                widgetType: widgetMeta.type,
-                widgetData: widgetData,
-                props: widgetData, // Also as 'props' for compatibility
-              }
-            }
-          };
-
-          // Record successful tool call
-          mcpHealthService.recordSuccess(serverId, mcpTool.name);
-
-          const structuredResult = {
-            text: `[UI Widget: ${widgetMeta.name}]`,
-            content: result.content,
-            uiResources: [uiResource],
-            _meta: result._meta,
-            structuredContent: result.structuredContent,
-          };
-
-          logger.aiSdk.info("[AI-SDK] Returning mcp-use widget result", {
-            toolName: mcpTool.name,
-            widgetName: widgetMeta.name,
-            htmlLength: widgetHtml.length,
-            usedServerHtml: !!widgetMeta.html,
-          });
-
-          return structuredResult;
+          return handleMcpUseWidget(serverId, mcpTool, args, result, widgetMeta);
         } else if (openaiOutputTemplate && typeof openaiOutputTemplate === 'string' && openaiOutputTemplate.startsWith('ui://')) {
-          // Handle Skybridge/OpenAI widget format
-          logger.aiSdk.info("[AI-SDK] Found openai/outputTemplate widget", {
-            toolName: mcpTool.name,
-            templateUri: openaiOutputTemplate,
-            hasMeta: !!result._meta,
-            hasStructuredContent: !!result.structuredContent,
-          });
-
-          try {
-            // Fetch widget HTML from the UI resource
-            const resourceContent = await mcpService.readResource(serverId, openaiOutputTemplate);
-
-            logger.aiSdk.info("[AI-SDK] Fetched Skybridge widget content", {
-              uri: openaiOutputTemplate,
-              contentsCount: resourceContent?.contents?.length || 0,
-              firstContentMimeType: resourceContent?.contents?.[0]?.mimeType,
-              firstContentHasText: !!resourceContent?.contents?.[0]?.text,
-            });
-
-            if (resourceContent?.contents?.[0]?.text) {
-              let widgetHtml = resourceContent.contents[0].text;
-
-              // Extract widget data from result._meta (excluding the outputTemplate key)
-              const widgetData: Record<string, any> = {};
-              if (result._meta) {
-                for (const [key, value] of Object.entries(result._meta)) {
-                  if (key !== 'openai/outputTemplate' && !key.startsWith('openai/')) {
-                    widgetData[key] = value;
-                  }
-                }
-              }
-
-              // Also include structuredContent if available
-              if (result.structuredContent) {
-                Object.assign(widgetData, result.structuredContent);
-              }
-
-              // Inject Skybridge/OpenAI bridge into widget HTML
-              // This provides window.openai with all the APIs that Skybridge widgets expect
-              widgetHtml = injectSkybridgeBridge(widgetHtml, {
-                toolInput: args,
-                toolOutput: result.structuredContent || widgetData,
-                responseMetadata: widgetData,
-                locale: typeof navigator !== 'undefined' ? navigator.language : 'en-US',
-              });
-
-              logger.aiSdk.info("[AI-SDK] Injected Skybridge bridge into widget HTML", {
-                toolName: mcpTool.name,
-                htmlLengthBefore: resourceContent.contents[0].text.length,
-                htmlLengthAfter: widgetHtml.length,
-                widgetDataKeys: Object.keys(widgetData),
-              });
-
-              // Normalize mimeType - UIResourceRenderer expects standard types
-              let mimeType = resourceContent.contents[0].mimeType || "text/html";
-              // Convert Skybridge-specific mimeTypes to standard ones
-              if (mimeType.startsWith('text/html+')) {
-                mimeType = 'text/html';
-              }
-
-              // Create UI resource with Skybridge widget
-              const uiResource = {
-                type: "resource",
-                resource: {
-                  uri: openaiOutputTemplate,
-                  mimeType: mimeType,
-                  text: widgetHtml,
-                  _meta: {
-                    widgetData: widgetData,
-                    props: widgetData,
-                    isSkybridge: true, // Mark as Skybridge widget for UIResourceMessage
-                  }
-                }
-              };
-
-              mcpHealthService.recordSuccess(serverId, mcpTool.name);
-
-              const structuredResult = {
-                text: `[UI Widget from ${openaiOutputTemplate}]`,
-                content: result.content,
-                uiResources: [uiResource],
-                _meta: result._meta,
-                structuredContent: result.structuredContent,
-              };
-
-              logger.aiSdk.info("[AI-SDK] Returning Skybridge widget result", {
-                toolName: mcpTool.name,
-                templateUri: openaiOutputTemplate,
-                htmlLength: widgetHtml.length,
-                widgetDataKeys: Object.keys(widgetData),
-              });
-
-              return structuredResult;
-            }
-          } catch (error) {
-            logger.aiSdk.error("[AI-SDK] Failed to fetch Skybridge widget", {
-              toolName: mcpTool.name,
-              templateUri: openaiOutputTemplate,
-              error: error instanceof Error ? error.message : String(error),
-            });
-            // Fall through to normal processing
-          }
+          const skybridgeResult = await handleSkybridgeWidget(serverId, mcpTool, args, result, openaiOutputTemplate);
+          if (skybridgeResult) return skybridgeResult;
+          // Fall through to normal processing if Skybridge handling fails
         }
 
         // Process MCP result - preserve UI resources for rendering
-        if (result.content && Array.isArray(result.content)) {
-          const textParts: string[] = [];
-          const uiResources: any[] = [];
-
-          for (const item of result.content) {
-            logger.aiSdk.debug("[AI-SDK] Processing content item", {
-              itemType: item.type,
-              hasText: !!item.text,
-              hasResource: !!item.resource,
-              hasData: !!item.data,
-              fullItem: JSON.stringify(item, null, 2)
-            });
-
-            if (item.type === "text") {
-              textParts.push(item.text || "");
-            } else if (item.type === "resource") {
-              // Check if this is a UI resource (uri starts with ui://)
-              let resourceData = item.resource || item.data || item;
-              const uri = resourceData?.uri || "";
-
-              logger.aiSdk.info("[AI-SDK] Found resource in tool result", {
-                uri,
-                isUIResource: uri.startsWith("ui://"),
-                mimeType: resourceData?.mimeType,
-                hasText: !!resourceData?.text,
-                hasBlob: !!resourceData?.blob,
-                textLength: resourceData?.text?.length || 0,
-                resourceKeys: Object.keys(resourceData || {})
-              });
-
-              if (uri.startsWith("ui://")) {
-                // If UI resource has no content, fetch it via readResource
-                if (!resourceData?.text && !resourceData?.blob) {
-                  logger.aiSdk.info("[AI-SDK] UI resource has no content, fetching via readResource", {
-                    serverId,
-                    uri
-                  });
-
-                  try {
-                    const resourceContent = await mcpService.readResource(serverId, uri);
-                    logger.aiSdk.info("[AI-SDK] Fetched UI resource content", {
-                      uri,
-                      contentsCount: resourceContent?.contents?.length || 0,
-                      firstContentMimeType: resourceContent?.contents?.[0]?.mimeType,
-                      firstContentHasText: !!resourceContent?.contents?.[0]?.text,
-                      firstContentTextLength: resourceContent?.contents?.[0]?.text?.length || 0
-                    });
-
-                    // Merge fetched content into resourceData
-                    if (resourceContent?.contents?.[0]) {
-                      const fetchedContent = resourceContent.contents[0];
-                      resourceData = {
-                        ...resourceData,
-                        uri: fetchedContent.uri || uri,
-                        mimeType: fetchedContent.mimeType || resourceData?.mimeType,
-                        text: fetchedContent.text,
-                        blob: fetchedContent.blob,
-                      };
-                    }
-                  } catch (fetchError) {
-                    logger.aiSdk.error("[AI-SDK] Failed to fetch UI resource content", {
-                      serverId,
-                      uri,
-                      error: fetchError instanceof Error ? fetchError.message : fetchError
-                    });
-                  }
-                }
-
-                // Preserve UI resource structure for rendering
-                uiResources.push({
-                  type: "resource",
-                  resource: resourceData
-                });
-                // Also add text description for the model
-                textParts.push(`[UI Widget: ${uri}]`);
-              } else {
-                // Regular resource - stringify for model
-                textParts.push(`[Resource: ${JSON.stringify(resourceData)}]`);
-              }
-            } else {
-              // Other types - stringify
-              logger.aiSdk.debug("[AI-SDK] Other content type", { type: item.type });
-              textParts.push(`[${item.type}: ${JSON.stringify(item.data || item)}]`);
-            }
-          }
-
-          // Record successful tool call
-          mcpHealthService.recordSuccess(serverId, mcpTool.name);
-
-          // Return structured result with both text and UI resources
-          // AI SDK v5 supports structured outputs
-          if (uiResources.length > 0) {
-            const structuredResult = {
-              text: textParts.join("\n"),
-              content: result.content,  // Preserve original content
-              uiResources: uiResources  // Extracted UI resources
-            };
-
-            logger.aiSdk.info("[AI-SDK] Returning structured result with UI resources", {
-              toolName: mcpTool.name,
-              uiResourceCount: uiResources.length,
-              textPartsCount: textParts.length,
-              uiResourceUris: uiResources.map(r => r.resource?.uri),
-              firstResourceHasText: !!uiResources[0]?.resource?.text,
-              firstResourceHasBlob: !!uiResources[0]?.resource?.blob,
-              structuredResultKeys: Object.keys(structuredResult)
-            });
-
-            return structuredResult;
-          }
-
-          // No UI resources - return text only
-          const resultText = textParts.join("\n");
-          logger.aiSdk.debug("Converted result text", { resultText });
-          return resultText;
-        }
-
-        // For non-content results, return as-is (structured)
-        logger.aiSdk.debug("Returning structured result", { result });
-
-        // Record successful tool call
-        mcpHealthService.recordSuccess(serverId, mcpTool.name);
-
-        return result;
+        return processToolResult(serverId, mcpTool, result);
       } catch (error) {
         const errorMessage = error instanceof Error ? error.message : "Tool execution failed";
 
@@ -1389,8 +260,7 @@ function createAISDKTool(serverId: string, mcpTool: Tool) {
         // Record failed tool call
         mcpHealthService.recordError(serverId, mcpTool.name, errorMessage);
 
-        // For tool execution errors, we should throw to let the AI SDK handle it
-        // This will trigger the 'tool-error' event in the stream
+        // For tool execution errors, throw to let the AI SDK handle it
         throw new Error(errorMessage);
       }
     },
@@ -1402,4 +272,229 @@ function createAISDKTool(serverId: string, mcpTool: Tool) {
   });
 
   return aiTool;
+}
+
+/**
+ * Handle mcp-use widget results
+ */
+function handleMcpUseWidget(
+  serverId: string,
+  mcpTool: Tool,
+  args: any,
+  result: any,
+  widgetMeta: any
+) {
+  logger.aiSdk.info("[AI-SDK] Found mcp-use widget in _meta", {
+    toolName: mcpTool.name,
+    widgetName: widgetMeta.name,
+    widgetType: widgetMeta.type,
+    hasHtml: !!widgetMeta.html,
+  });
+
+  // Use HTML from server if available, otherwise generate fallback
+  let widgetHtml: string;
+  const widgetData = result.structuredContent || args;
+
+  if (widgetMeta.html && typeof widgetMeta.html === 'string') {
+    // Server provided the HTML template - inject the data into it
+    widgetHtml = injectDataIntoHtml(widgetMeta.html, widgetData);
+  } else {
+    // Fallback: generate HTML ourselves
+    widgetHtml = generateWidgetHtml(widgetMeta, widgetData);
+  }
+
+  // Create synthetic UI resource with widget data in _meta
+  const uiResource = {
+    type: "resource",
+    resource: {
+      uri: `ui://widget/${widgetMeta.name}.html`,
+      mimeType: "text/html",
+      text: widgetHtml,
+      _meta: {
+        widgetName: widgetMeta.name,
+        widgetType: widgetMeta.type,
+        widgetData: widgetData,
+        props: widgetData,
+      }
+    }
+  };
+
+  // Record successful tool call
+  mcpHealthService.recordSuccess(serverId, mcpTool.name);
+
+  return {
+    text: `[UI Widget: ${widgetMeta.name}]`,
+    content: result.content,
+    uiResources: [uiResource],
+    _meta: result._meta,
+    structuredContent: result.structuredContent,
+  };
+}
+
+/**
+ * Handle Skybridge/OpenAI widget results
+ */
+async function handleSkybridgeWidget(
+  serverId: string,
+  mcpTool: Tool,
+  args: any,
+  result: any,
+  openaiOutputTemplate: string
+): Promise<any | null> {
+  logger.aiSdk.info("[AI-SDK] Found openai/outputTemplate widget", {
+    toolName: mcpTool.name,
+    templateUri: openaiOutputTemplate,
+  });
+
+  try {
+    // Fetch widget HTML from the UI resource
+    const resourceContent = await mcpService.readResource(serverId, openaiOutputTemplate);
+
+    if (resourceContent?.contents?.[0]?.text) {
+      let widgetHtml = resourceContent.contents[0].text;
+
+      // Extract widget data from result._meta (excluding the outputTemplate key)
+      const widgetData: Record<string, any> = {};
+      if (result._meta) {
+        for (const [key, value] of Object.entries(result._meta)) {
+          if (key !== 'openai/outputTemplate' && !key.startsWith('openai/')) {
+            widgetData[key] = value;
+          }
+        }
+      }
+
+      // Also include structuredContent if available
+      if (result.structuredContent) {
+        Object.assign(widgetData, result.structuredContent);
+      }
+
+      // Inject Skybridge/OpenAI bridge into widget HTML
+      widgetHtml = injectSkybridgeBridge(widgetHtml, {
+        toolInput: args,
+        toolOutput: result.structuredContent || widgetData,
+        responseMetadata: widgetData,
+        locale: 'en-US',
+      });
+
+      // Normalize mimeType
+      let mimeType = resourceContent.contents[0].mimeType || "text/html";
+      if (mimeType.startsWith('text/html+')) {
+        mimeType = 'text/html';
+      }
+
+      // Create UI resource with Skybridge widget
+      const uiResource = {
+        type: "resource",
+        resource: {
+          uri: openaiOutputTemplate,
+          mimeType: mimeType,
+          text: widgetHtml,
+          _meta: {
+            widgetData: widgetData,
+            props: widgetData,
+            isSkybridge: true,
+          }
+        }
+      };
+
+      mcpHealthService.recordSuccess(serverId, mcpTool.name);
+
+      return {
+        text: `[UI Widget from ${openaiOutputTemplate}]`,
+        content: result.content,
+        uiResources: [uiResource],
+        _meta: result._meta,
+        structuredContent: result.structuredContent,
+      };
+    }
+  } catch (error) {
+    logger.aiSdk.error("[AI-SDK] Failed to fetch Skybridge widget", {
+      toolName: mcpTool.name,
+      templateUri: openaiOutputTemplate,
+      error: error instanceof Error ? error.message : String(error),
+    });
+  }
+
+  return null;
+}
+
+/**
+ * Process standard MCP tool results
+ */
+async function processToolResult(
+  serverId: string,
+  mcpTool: Tool,
+  result: any
+) {
+  if (result.content && Array.isArray(result.content)) {
+    const textParts: string[] = [];
+    const uiResources: any[] = [];
+
+    for (const item of result.content) {
+      if (item.type === "text") {
+        textParts.push(item.text || "");
+      } else if (item.type === "resource") {
+        // Check if this is a UI resource (uri starts with ui://)
+        let resourceData = item.resource || item.data || item;
+        const uri = resourceData?.uri || "";
+
+        if (uri.startsWith("ui://")) {
+          // If UI resource has no content, fetch it via readResource
+          if (!resourceData?.text && !resourceData?.blob) {
+            try {
+              const resourceContent = await mcpService.readResource(serverId, uri);
+              if (resourceContent?.contents?.[0]) {
+                const fetchedContent = resourceContent.contents[0];
+                resourceData = {
+                  ...resourceData,
+                  uri: fetchedContent.uri || uri,
+                  mimeType: fetchedContent.mimeType || resourceData?.mimeType,
+                  text: fetchedContent.text,
+                  blob: fetchedContent.blob,
+                };
+              }
+            } catch (fetchError) {
+              logger.aiSdk.error("[AI-SDK] Failed to fetch UI resource content", {
+                serverId,
+                uri,
+                error: fetchError instanceof Error ? fetchError.message : fetchError
+              });
+            }
+          }
+
+          // Preserve UI resource structure for rendering
+          uiResources.push({
+            type: "resource",
+            resource: resourceData
+          });
+          textParts.push(`[UI Widget: ${uri}]`);
+        } else {
+          // Regular resource - stringify for model
+          textParts.push(`[Resource: ${JSON.stringify(resourceData)}]`);
+        }
+      } else {
+        // Other types - stringify
+        textParts.push(`[${item.type}: ${JSON.stringify(item.data || item)}]`);
+      }
+    }
+
+    // Record successful tool call
+    mcpHealthService.recordSuccess(serverId, mcpTool.name);
+
+    // Return structured result with both text and UI resources
+    if (uiResources.length > 0) {
+      return {
+        text: textParts.join("\n"),
+        content: result.content,
+        uiResources: uiResources
+      };
+    }
+
+    // No UI resources - return text only
+    return textParts.join("\n");
+  }
+
+  // For non-content results, return as-is
+  mcpHealthService.recordSuccess(serverId, mcpTool.name);
+  return result;
 }
