+++ src/main/services/ai/widgets/templates/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Widget Templates
+ *
+ * Pre-built HTML templates for common widget types.
+ */
+
+export { generateWeatherCardHtml, getWeatherEmoji } from './weatherCard';
+export { generateCalculatorResultHtml } from './calculator';
+export { generateTextAnalysisChartHtml } from './textAnalysis';
+export { generateGenericWidgetHtml } from './generic';
